pyproject.toml
[project]
name = "scheme-generator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "lxml>=6.1.0",
    "pydantic>=2.13.4",
    "pyyaml>=6.0.3",
]
